refactor: mix architecture change single (#37)
