+++ api/src/main/kotlin/team/backend/common/entity/Subject.kt
@@ -1,0 +1,30 @@
+package team.backend.common.entity
+
+import jakarta.persistence.Column
+import jakarta.persistence.Entity
+import jakarta.persistence.EnumType
+import jakarta.persistence.Enumerated
+import jakarta.persistence.GeneratedValue
+import jakarta.persistence.GenerationType
+import jakarta.persistence.Id
+import jakarta.persistence.Table
+
+@Entity
+@Table(name = "tbl_subject")
+class Subject(
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    @Column(name = "subject_id")
+    val id: Long? = null,
+
+    @Enumerated(EnumType.STRING)
+    @Column(name = "subject_type", nullable = false, insertable = true, updatable = false)
+    val type: SubjectType
+): BaseEntity()
+
+enum class SubjectType {
+    SPRING,
+    JAVA,
+    NETWORK,
+    OS,
+}
