+++ api/src/main/kotlin/team/backend/domain/repository/member/MemberQuizAnswerStore.kt
@@ -1,0 +1,8 @@
+package team.backend.domain.repository.member
+
+import team.backend.domain.entity.member.MemberQuizAnswer
+
+interface MemberQuizAnswerStore {
+
+    suspend fun save(entity: MemberQuizAnswer)
+}
