+++ api/src/main/kotlin/team/backend/domain/repository/member/MemberStore.kt
@@ -1,0 +1,4 @@
+package team.backend.domain.repository.member
+
+interface MemberStore {
+}
