+++ api/src/main/kotlin/team/backend/domain/repository/prompt/PromptStore.kt
@@ -1,0 +1,4 @@
+package team.backend.domain.repository.prompt
+
+interface PromptStore {
+}
