+++ api/src/main/kotlin/team/backend/domain/repository/quiz/QuizStore.kt
@@ -1,0 +1,4 @@
+package team.backend.domain.repository.quiz
+
+interface QuizStore {
+}
