+++ api/src/main/kotlin/team/backend/infrastructure/service/quiz/QuizService.kt
@@ -1,0 +1,8 @@
+package team.backend.infrastructure.service.quiz
+
+import team.backend.domain.command.quiz.QuizCommand
+
+interface QuizService {
+
+    suspend fun answer(command: QuizCommand.AnswerRequest): QuizCommand.AnswerResponse
+}
