--- api/src/main/kotlin/team/backend/member/domain/repository/MemberQuizAnswerStore.kt
@@ -1,8 +1,0 @@
-package team.backend.member.domain.repository
-
-import team.backend.member.domain.entity.MemberQuizAnswer
-
-interface MemberQuizAnswerStore {
-
-    suspend fun save(entity: MemberQuizAnswer)
-}
