--- api/src/main/kotlin/team/backend/member/domain/repository/MemberReader.kt
@@ -1,8 +1,0 @@
-package team.backend.member.domain.repository
-
-import team.backend.member.domain.entity.Member
-
-interface MemberReader {
-
-    suspend fun get(id: Long): Member
-}
