--- api/src/main/kotlin/team/backend/member/domain/repository/MemberStore.kt
@@ -1,4 +1,0 @@
-package team.backend.member.domain.repository
-
-interface MemberStore {
-}
