--- api/src/main/kotlin/team/backend/member/infrastructure/repository/MemberQuizAnswerStoreImpl.kt
@@ -1,21 +1,0 @@
-package team.backend.member.infrastructure.repository
-
-import com.linecorp.kotlinjdsl.spring.data.reactive.query.SpringDataHibernateMutinyReactiveQueryFactory
-import org.hibernate.reactive.mutiny.impl.MutinySessionFactoryImpl
-import org.springframework.stereotype.Repository
-import team.backend.member.domain.entity.MemberQuizAnswer
-import team.backend.member.domain.repository.MemberQuizAnswerStore
-
-@Repository
-class MemberQuizAnswerStoreImpl(
-    private val queryFactory: SpringDataHibernateMutinyReactiveQueryFactory,
-    private val mutinySessionFactory: MutinySessionFactoryImpl,
-): MemberQuizAnswerStore {
-
-    override suspend fun save(entity: MemberQuizAnswer) {
-        mutinySessionFactory.withSession { session ->
-            session.persist(entity)
-                .flatMap { session.flush() } }
-            .subscribeAsCompletionStage()
-    }
-}
