+++ api/src/main/kotlin/team/backend/presentation/quiz/dto/QuizDto.kt
@@ -1,0 +1,19 @@
+package team.backend.presentation.quiz.dto
+
+import jakarta.validation.constraints.NotEmpty
+import jakarta.validation.constraints.NotNull
+
+class QuizDto {
+
+    data class AnswerRequest(
+        @field:NotNull
+        val quizId: Long,
+        @field:NotEmpty(message = "Blank spaces are not possible.")
+        val answer: String,
+    )
+
+    data class AnswerResponse(
+        val isCorrect: Boolean,
+        val answer: String,
+    )
+}
