--- api/src/main/kotlin/team/backend/prompt/domain/repository/PromptStore.kt
@@ -1,4 +1,0 @@
-package team.backend.prompt.domain.repository
-
-interface PromptStore {
-}
