--- api/src/main/kotlin/team/backend/quiz/domain/repository/QuizReader.kt
@@ -1,8 +1,0 @@
-package team.backend.quiz.domain.repository
-
-import team.backend.quiz.domain.entity.Quiz
-
-interface QuizReader {
-
-    suspend fun get(id: Long): Quiz
-}
